All Expense functions completed

+++ src/main/java/com/WTT/ExpenseTrackingAppBE/services/ExpenseServiceImpl.java
@@ -2,6 +2,7 @@ package com.WTT.ExpenseTrackingAppBE.services;
 
 import com.WTT.ExpenseTrackingAppBE.dto.ExpenseDto;
 import com.WTT.ExpenseTrackingAppBE.repos.ExpenseRepo;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +11,10 @@ import org.springframework.stereotype.Service;
 
 import com.WTT.ExpenseTrackingAppBE.entities.Expense;
 
+import java.util.Comparator;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 
 @Service
@@ -18,7 +22,7 @@ import java.util.List;
 public class ExpenseServiceImpl implements ExpenseService{
 
     @Autowired
-    private final ExpenseRepo expenseRepo;
+    private ExpenseRepo expenseRepo;
     /**
      * @param expenseDto
      * @return
@@ -33,7 +37,9 @@ public class ExpenseServiceImpl implements ExpenseService{
      */
     @Override
     public List<Expense> getAllExpenses() {
-        return expenseRepo.findAll();
+        return expenseRepo.findAll().stream()
+                .sorted(Comparator.comparing(Expense::getDate).reversed())
+                .collect(Collectors.toList());
     }
 
     /**
@@ -60,7 +66,32 @@ public class ExpenseServiceImpl implements ExpenseService{
      */
     @Override
     public Expense getExpenseById(Long id) {
-        return expenseRepo.findById(id).orElse(null);
+        Optional<Expense> optionalExpense = expenseRepo.findById(id);
+        if(optionalExpense.isPresent()){
+            return optionalExpense.get();
+
+        }else {
+            throw new EntityNotFoundException("Expense with id "+id
+                    +" was not found");
+        }
+
+    }
+
+    /**
+     * @param id
+     * @param expenseDto
+     * @return
+     */
+    @Override
+    public Expense updateExpense(Long id, ExpenseDto expenseDto) {
+        Optional<Expense> optionalExpense = expenseRepo.findById(id);
+        if(optionalExpense.isPresent()){
+            return saveExpense(optionalExpense.get(), expenseDto);
+
+        }else {
+            throw new EntityNotFoundException("Expense with id "+id
+                    +" was not found");
+        }
     }
 
     /**
@@ -68,7 +99,13 @@ public class ExpenseServiceImpl implements ExpenseService{
      */
     @Override
     public void deleteExpense(Long id) {
-        expenseRepo.deleteById(id);
+        Optional<Expense> optionalExpense = expenseRepo.findById(id);
+        if(optionalExpense.isPresent()){
+            expenseRepo.deleteById(id);
+        }else {
+            throw new EntityNotFoundException("Expense with id "+id
+                    +" was not found");
+        }
 
     }
 }
